Move li_dev4's content to li_dev5, including no-sorting read-format, and overflow of fragment length

+++ src/mapping_in_memory.h
@@ -83,7 +83,7 @@ struct PairedEndMappingInMemory {
     return mapping_in_memory2.GetFragmentStartPosition();
   }
 
-  inline uint16_t GetFragmentLength() const {
+  inline int GetFragmentLength() const {
     if (mapping_in_memory1.strand == kPositive) {
       return mapping_in_memory2.ref_end_position -
              mapping_in_memory1.ref_start_position + 1;
